requirements.txt
pymongo == 3.4.0
pyserial
py-sds011==0.9
